Break code out into multiple files
This also adds a lot of TODO comments for things to do later
