+++ src/main.rs
@@ -11,27 +11,25 @@ use std::path::{Path, PathBuf};
 use std::sync::mpsc;
 
 use std::{env, fmt, io, iter, process, str, string, thread, time};
+
+mod bundler;
 mod es6;
+mod input_options;
+mod manifest;
+mod modules;
 mod opts;
+mod path_ext;
+mod resolver;
+mod source_maps;
+mod vlq;
+mod worker;
+mod writer;
 
-macro_rules! map {
-    {} => {
-        Default::default()
-    };
-    { $($key:expr => $value:expr,)+ } => {
-        {
-            let mut map = FnvHashMap::default();
-            $(map.insert($key, $value);)+
-            map
-        }
-    };
-    { $($key:expr => $value:expr),+ } => {
-        map!{$($key => $value,)+}
-    };
-}
+use bundler::bundle;
+use input_options::{InputOptions, PackageManager};
+use resolver::{Resolved, Resolver};
+use source_maps::SourceMapOutput;
 
-const HEAD_JS: &str = include_str!("head.js");
-const TAIL_JS: &str = include_str!("tail.js");
 const CORE_MODULES: &[&str] = &[
     "assert",
     "buffer",
@@ -62,657 +60,14 @@ const CORE_MODULES: &[&str] = &[
     "zlib",
 ];
 
-fn cjs_parse_deps<'f, 's>(
-    lex: &mut lex::Lexer<'f, 's>,
-) -> Result<FnvHashSet<Cow<'s, str>>, CliError> {
-    // TODO should we panic on dynamic requires?
-    let mut deps = FnvHashSet::default();
-    loop {
-        eat!(lex,
-            // Tt::Id(s) if s == "require" => eat!(lex,
-            Tt::Id("require") => eat!(lex,
-                Tt::Lparen => eat!(lex,
-                    Tt::StrLitSgl(s) |
-                    Tt::StrLitDbl(s) => eat!(lex,
-                        Tt::Rparen => {
-                            deps.insert(lex::str_lit_value(s)?);
-                        },
-                        _ => {},
-                    ),
-                    _ => {},
-                ),
-                // Tt::Dot => eat!(lex,
-                //     Tt::Id("resolve") => eat!(lex,
-                //         Tt::Lparen => eat!(lex,
-                //             Tt::StrLitSgl(s) |
-                //             Tt::StrLitDbl(s) => eat!(lex,
-                //                 Tt::Rparen => {
-                //                     // TODO handle error
-                //                     return Some(lex::str_lit_value(s).unwrap())
-                //                 },
-                //                 _ => {},
-                //             ),
-                //             _ => {},
-                //         ),
-                //         _ => {},
-                //     ),
-                //     _ => {},
-                // ),
-                _ => {},
-            ),
-            Tt::Eof => return Ok(deps),
-            _ => {
-                lex.advance();
-            },
-        );
-    }
-}
-
-#[derive(Debug)]
-struct Writer<'a, 'b> {
-    modules: FnvHashMap<PathBuf, Module>,
-    entry_point: &'a Path,
-    map_output: &'b SourceMapOutput<'b>,
-}
-
-impl<'a, 'b> Writer<'a, 'b> {
-    fn sorted_modules(&self) -> Vec<(&Path, &Module)> {
-        let mut modules = self
-            .modules
-            .iter()
-            .map(|(p, m)| (p.as_path(), m))
-            .collect::<Vec<_>>();
-        modules.sort_by(|(f, _), (g, _)| f.cmp(g));
-        modules
-    }
-
-    fn write_to<W: io::Write>(&self, w: &mut W) -> io::Result<()> {
-        w.write_all(HEAD_JS.as_bytes())?;
-        // for (module, main) in self.mains {
-        //     write!(w,
-        //         "\n  Pax.mains[{mod_path}] = {main_path}",
-        //         mod_path = Self::js_path(&module),
-        //         main_path = Self::js_path(&main),
-        //     );
-        // }
-
-        for (file, info) in self.sorted_modules() {
-            let id = Self::name_path(&file);
-            let deps = Self::stringify_deps(&info.deps, PathBuf::from(self.entry_point));
-            let filename = Self::js_path(&file);
-
-            write!(w,
-                "\n  Pax.files[{filename}] = {id}; {id}.deps = {deps}; {id}.filename = {filename}; function {id}(module, exports, require, __filename, __dirname, __import_meta) {{\n",
-                filename = filename,
-                id = id,
-                deps = deps,
-            )?;
-            if !info.source.prefix.is_empty() {
-                w.write_all(info.source.prefix.as_bytes())?;
-                w.write_all(b"\n")?;
-            }
-            w.write_all(info.source.body.as_bytes())?;
-            if !matches!(
-                info.source.body.chars().last(),
-                None | Some('\n') | Some('\r') | Some('\u{2028}') | Some('\u{2029}')
-            ) {
-                w.write_all(b"\n")?;
-            }
-            if !info.source.suffix.is_empty() {
-                w.write_all(info.source.suffix.as_bytes())?;
-            }
-            write!(w, "}}")?;
-        }
-        let main = Self::name_path(
-            self.entry_point
-                .strip_prefix(self.entry_point.parent().unwrap())
-                .unwrap(),
-        );
-        write!(w,
-            "\n  Pax.main = {main}; Pax.makeRequire(null)()\n  if (typeof module !== 'undefined') module.exports = Pax.main.module && Pax.main.module.exports\n",
-            main = main,
-        )?;
-        w.write_all(TAIL_JS.as_bytes())?;
-        match *self.map_output {
-            SourceMapOutput::Suppressed => {}
-            SourceMapOutput::Inline => {
-                let mut map = Vec::new();
-                self.write_map_to(&mut map)?;
-                writeln!(
-                    w,
-                    "//# sourceMappingURL=data:application/json;charset=utf-8;base64,{data}",
-                    data = base64::encode(&map),
-                )?;
-            }
-            SourceMapOutput::File(ref path, output_file) => {
-                // TODO handle error
-                let relative = path.relative_from(output_file.parent().unwrap());
-                let map = relative.as_ref().unwrap_or(path);
-                writeln!(w, "//# sourceMappingURL={map}", map = map.display(),)?;
-            }
-        }
-        Ok(())
-    }
-
-    fn write_map_to<W: io::Write>(&self, w: &mut W) -> serde_json::Result<()> {
-        // https://sourcemaps.info/spec.html
-
-        let modules = &self.sorted_modules();
-        let dir = self.entry_point.parent().unwrap();
-
-        #[derive(Serialize, Debug)]
-        #[serde(rename_all = "camelCase")]
-        struct SourceMap<'a> {
-            version: u8,
-            file: &'static str,
-            source_root: &'static str,
-            sources: Sources<'a>,
-            sources_content: SourcesContent<'a>,
-            names: [(); 0],
-            mappings: Mappings<'a>,
-        }
-
-        #[derive(Debug)]
-        struct Sources<'a> {
-            modules: &'a [(&'a Path, &'a Module)],
-            dir: &'a Path,
-        }
-
-        impl<'a> Serialize for Sources<'a> {
-            fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-                let mut seq = serializer.serialize_seq(None)?;
-                for (f, _) in self.modules {
-                    let relative = f.relative_from(self.dir);
-                    let path = relative.as_ref().map_or(*f, PathBuf::as_path);
-                    seq.serialize_element(&path.to_string_lossy())?;
-                }
-                seq.end()
-            }
-        }
-
-        #[derive(Debug)]
-        struct SourcesContent<'a> {
-            modules: &'a [(&'a Path, &'a Module)],
-        }
-
-        impl<'a> Serialize for SourcesContent<'a> {
-            fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-                let mut seq = serializer.serialize_seq(None)?;
-                for (_, module) in self.modules {
-                    let content = module
-                        .source
-                        .original
-                        .as_ref()
-                        .unwrap_or(&module.source.body);
-                    seq.serialize_element(content)?;
-                }
-                seq.end()
-            }
-        }
-
-        #[derive(Debug)]
-        struct Mappings<'a> {
-            modules: &'a [(&'a Path, &'a Module)],
-        }
-
-        impl<'a> Serialize for Mappings<'a> {
-            fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-                serializer.collect_str(self)
-            }
-        }
-
-        impl<'a> Display for Mappings<'a> {
-            fn fmt(&self, w: &mut fmt::Formatter) -> fmt::Result {
-                let prefix_len = count_lines(HEAD_JS); /*+ this.mains.size*/
-                for _ in 0..prefix_len {
-                    w.write_str(";")?;
-                }
-
-                let mut line = 0;
-                let mut vlq = Vlq::new();
-                for (index, &(_, module)) in self.modules.iter().enumerate() {
-                    w.write_str(";")?;
-                    if !module.source.prefix.is_empty() {
-                        for _ in 0..count_lines(&module.source.prefix) {
-                            w.write_str(";")?;
-                        }
-                    }
-                    for i in 0..count_lines(&module.source.body) {
-                        w.write_str("A")?;
-                        if i == 0 {
-                            if index == 0 {
-                                w.write_str("AAA")?;
-                            } else {
-                                w.write_str("C")?;
-                                w.write_str(vlq.enc(-line))?;
-                                w.write_str("A")?;
-                            }
-                            line = 0;
-                        } else {
-                            w.write_str("ACA")?;
-                            line += 1;
-                        }
-                        w.write_str(";")?;
-                    }
-                    if !matches!(
-                        module.source.body.chars().last(),
-                        None | Some('\n') | Some('\r') | Some('\u{2028}') | Some('\u{2029}')
-                    ) {
-                        w.write_str(";")?;
-                    }
-                    for _ in 0..count_lines(&module.source.suffix) - 1 {
-                        w.write_str(";")?;
-                    }
-                }
-                for _ in 0..2 + count_lines(TAIL_JS) + 1 - 1 - 1 {
-                    w.write_str(";")?;
-                }
-                Ok(())
-            }
-        }
-
-        serde_json::to_writer(
-            w,
-            &SourceMap {
-                version: 3,
-                file: "",
-                source_root: "",
-                sources: Sources { modules, dir },
-                sources_content: SourcesContent { modules },
-                names: [],
-                mappings: Mappings { modules },
-            },
-        )
-    }
-
-    fn stringify_deps(deps: &FnvHashMap<String, Resolved>, entry_point: PathBuf) -> String {
-        let mut result = "{".to_owned();
-        let mut comma = false;
-        for (name, resolved) in deps {
-            match *resolved {
-                Resolved::External => {}
-                Resolved::Ignore => {
-                    if comma {
-                        result.push(',');
-                    }
-                    result.push_str(&to_quoted_json_string(name));
-                    result.push_str(":Pax.ignored");
-                    comma = true;
-                }
-                Resolved::Normal(ref path) => {
-                    if comma {
-                        result.push(',');
-                    }
-                    result.push_str(&to_quoted_json_string(name));
-                    result.push(':');
-                    let parent = entry_point.parent().unwrap();
-
-                    match path.as_path().strip_prefix(parent) {
-                        Ok(path) => Self::write_name_path(path, &mut result),
-                        Err(_) => {
-                            Self::write_name_path(path, &mut result);
-                        }
-                    }
-                    comma = true;
-                }
-            }
-        }
-        result.push('}');
-        result
-    }
-
-    #[cfg(target_os = "windows")]
-    fn js_path(path: &Path) -> String {
-        // TODO untested
-        let string = path.to_string_lossy();
-        let replaced = string.replace('\\', "/");
-        to_quoted_json_string(&replaced)
-    }
-
-    #[cfg(not(target_os = "windows"))]
-    fn js_path(path: &Path) -> String {
-        let string = path.to_string_lossy();
-        to_quoted_json_string(&string)
-    }
-
-    fn name_path(path: &Path) -> String {
-        let mut result = String::new();
-        Self::write_name_path(path, &mut result);
-        result
-    }
-    fn write_name_path(path: &Path, result: &mut String) {
-        let string = path.to_string_lossy();
-        // let slice = string.as_ref();
-        let bytes = string.as_bytes();
-
-        result.push_str("file_");
-        for &b in bytes {
-            match b {
-                b'_' | b'a'..=b'z' | b'A'..=b'Z' | b'0'..=b'9' => {
-                    result.push(b as char);
-                }
-                _ => {
-                    write!(result, "${:02x}", b).unwrap();
-                }
-            }
-        }
-
-        // let mut last_pos = 0;
-        // for pos in bytes.match_indices(|&b| {
-        //     !matches!(b as char, '_' | 'a'...'z' | 'A'...'Z')
-        // }) {
-        //     result.push_str(slice[last_pos..pos]);
-        //     write!(result, "${:02x}", bytes[pos]);
-        //     last_pos = pos + 1;
-        // }
-    }
-}
-
-fn to_quoted_json_string(s: &str) -> String {
-    // Serializing to a String only fails if the Serialize impl decides to fail, which the Serialize impl of `str` never does.
-    serde_json::to_string(s).unwrap()
-}
-
-fn count_lines(source: &str) -> usize {
+pub fn count_lines(source: &str) -> usize {
     // TODO non-ASCII line terminators?
     1 + memchr::Memchr::new(b'\n', source.as_bytes()).count()
 }
 
-struct Vlq {
-    buf: [u8; 13],
-}
-impl Vlq {
-    fn new() -> Self {
-        Self { buf: [0u8; 13] }
-    }
-
-    fn enc(&mut self, n: isize) -> &str {
-        let sign = n < 0;
-        let n = if sign { -n } else { n } as usize;
-        let mut y = (n & 0xf) << 1 | sign as usize;
-        let mut r = n >> 4;
-        let mut l = 0;
-        while r > 0 {
-            y |= 0x20;
-            self.buf[l] = B64[y];
-            y = r & 0x1f;
-            r >>= 5;
-            l += 1;
-        }
-        self.buf[l] = B64[y];
-        str::from_utf8(&self.buf[0..=l]).unwrap()
-    }
-}
-const B64: &[u8] = b"ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
-
-#[derive(Debug, Clone)]
-struct WorkerInit {
-    tx: mpsc::Sender<Result<WorkDone, CliError>>,
-    input_options: InputOptions,
-    queue: Arc<SegQueue<Work>>,
-    quit: Arc<AtomicBool>,
-}
-#[derive(Debug, Clone)]
-struct Worker {
-    tx: mpsc::Sender<Result<WorkDone, CliError>>,
-    resolver: Resolver,
-    queue: Arc<SegQueue<Work>>,
-    quit: Arc<AtomicBool>,
-}
-
-#[derive(Debug, Clone, Default)]
-struct Resolver {
-    input_options: InputOptions,
-    cache: PackageCache,
-}
-
-#[derive(Debug, Clone, Default)]
-struct PackageCache {
-    pkgs: RefCell<FnvHashMap<PathBuf, Option<Rc<PackageInfo>>>>,
-}
-
-#[derive(Debug)]
-enum Work {
-    Resolve { context: PathBuf, name: String },
-    Include { module: PathBuf },
-}
-#[derive(Debug)]
-enum WorkDone {
-    Resolve {
-        context: PathBuf,
-        name: String,
-        resolved: Resolved,
-    },
-    Include {
-        module: PathBuf,
-        info: ModuleInfo,
-    },
-}
-#[derive(Debug)]
-enum ModuleState {
-    Loading,
-    Loaded(Module),
-}
-#[derive(Debug)]
-pub struct Module {
-    pub source: Source,
-    pub deps: FnvHashMap<String, Resolved>,
-}
-#[derive(Debug)]
-struct ModuleInfo {
-    source: Source,
-    deps: Vec<String>,
-}
-#[derive(Debug)]
-pub struct Source {
-    pub prefix: String,
-    pub body: String,
-    pub suffix: String,
-    pub original: Option<String>,
-}
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum Resolved {
-    External,
-    Ignore,
-    // CoreWithSubst(PathBuf),
-    Normal(PathBuf),
-}
-#[derive(Debug, Clone, PartialEq, Eq)]
-enum ModuleSubstitution {
-    Normal,
-    Ignore,
-    External,
-    Replace(String),
-}
-#[derive(Debug, Clone, PartialEq, Eq)]
-enum PathSubstitution {
-    Missing,
-    Normal,
-    Ignore,
-    Replace(PathBuf),
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Default)]
-pub struct InputOptions {
-    pub for_browser: bool,
-    pub for_bower: bool,
-    pub es6_syntax: bool,
-    pub es6_syntax_everywhere: bool,
-    pub external: FnvHashSet<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum SourceMapOutput<'a> {
-    Suppressed,
-    Inline,
-    File(PathBuf, &'a Path),
-}
-
-impl ModuleState {
-    fn expect(self, message: &str) -> Module {
-        match self {
-            ModuleState::Loading => panic!("{}", message),
-            ModuleState::Loaded(module) => module,
-        }
-    }
-    fn unwrap(self) -> Module {
-        self.expect("unwrapped ModuleState that was still loading")
-    }
-}
-
-pub fn bundle(
-    entry_point: &Path,
-    input_options: InputOptions,
-    output: &str,
-    map_output: &SourceMapOutput,
-) -> Result<FnvHashMap<PathBuf, Module>, CliError> {
-    let mut pending = 0;
-    let thread_count = num_cpus::get();
-    let (tx, rx) = mpsc::channel();
-    let worker_init = WorkerInit {
-        tx,
-        input_options,
-        quit: Arc::new(AtomicBool::new(false)),
-        queue: Arc::new(SegQueue::new()),
-    };
-
-    // TODO: context.require('…')
-    // TODO: watch for missing files on error?
-
-    let mut modules = FnvHashMap::<PathBuf, ModuleState>::default();
-
-    worker_init.add_work(Work::Include {
-        module: entry_point.to_owned(),
-    });
-    pending += 1;
-    modules.insert(entry_point.to_owned(), ModuleState::Loading);
-
-    let children: Vec<_> = (0..thread_count)
-        .map(|_| {
-            let init = worker_init.clone();
-            thread::spawn(move || Worker::new(init).run())
-        })
-        .collect();
-    // let children: Vec<_> = (0..thread_count).map(|n| {
-    //     let init = worker_init.clone();
-    //     thread::Builder::new().name(format!("worker #{}", n + 1)).spawn(move || Worker::new(init).run()).unwrap()
-    // }).collect();
-
-    while let Ok(work_done) = rx.recv() {
-        // eprintln!("{:?}", work_done);
-        let work_done = match work_done {
-            Err(error) => {
-                worker_init.quit.store(true, Ordering::Relaxed);
-                return Err(error);
-            }
-            Ok(work_done) => {
-                pending -= 1;
-                work_done
-            }
-        };
-        match work_done {
-            WorkDone::Resolve {
-                context,
-                name,
-                resolved,
-            } => {
-                match *modules.get_mut(&context).unwrap() {
-                    ModuleState::Loading => unreachable!(),
-                    ModuleState::Loaded(Module { ref mut deps, .. }) => {
-                        deps.insert(name, resolved.clone());
-                    }
-                }
-                match resolved {
-                    Resolved::External => {}
-                    Resolved::Ignore => {}
-                    Resolved::Normal(module) => {
-                        modules.entry(module.clone()).or_insert_with(|| {
-                            worker_init.add_work(Work::Include { module });
-                            pending += 1;
-                            ModuleState::Loading
-                        });
-                    }
-                }
-            }
-            WorkDone::Include { module, info } => {
-                let old = modules.insert(
-                    module.clone(),
-                    ModuleState::Loaded(Module {
-                        source: info.source,
-                        deps: FnvHashMap::default(),
-                    }),
-                );
-                debug_assert_matches!(old, Some(ModuleState::Loading));
-                for dep in info.deps {
-                    worker_init.add_work(Work::Resolve {
-                        context: module.clone(),
-                        name: dep,
-                    });
-                    pending += 1;
-                }
-            }
-        }
-        if pending == 0 {
-            break;
-        }
-    }
-
-    worker_init.quit.store(true, Ordering::Relaxed);
-    for child in children {
-        child.join()?;
-    }
-
-    let writer = Writer {
-        modules: modules
-            .into_iter()
-            .map(|(k, ms)| {
-                let parent = entry_point.parent().unwrap();
-                let ms = ms.unwrap();
-
-                match k.as_path().strip_prefix(parent) {
-                    Ok(path) => (PathBuf::from(path), ms),
-                    Err(_) => (k, ms),
-                }
-            })
-            .collect(),
-        entry_point,
-        map_output,
-    };
-
-    match &*output {
-        "-" => {
-            let stdout = io::stdout();
-            let mut handle = stdout.lock();
-            writer.write_to(&mut handle)?;
-        }
-        _ => {
-            let output = Path::new(output);
-            if let Some(parent) = output.parent() {
-                fs::create_dir_all(parent)?;
-            }
-            let file = fs::File::create(&output)?;
-            let mut buf_writer = io::BufWriter::new(file);
-            writer.write_to(&mut buf_writer)?;
-        }
-    }
-    match *map_output {
-        SourceMapOutput::Suppressed => {}
-        SourceMapOutput::Inline => {
-            // handled in Writer::write_to()
-        }
-        SourceMapOutput::File(ref path, _) => {
-            if let Some(parent) = path.parent() {
-                fs::create_dir_all(parent)?;
-            }
-            let file = fs::File::create(path)?;
-            let mut buf_writer = io::BufWriter::new(file);
-            writer.write_map_to(&mut buf_writer)?;
-        }
-    }
-    // println!("entry point: {:?}", entry_point);
-    // println!("{:#?}", modules);
-
-    Ok(writer.modules)
+pub fn to_quoted_json_string(s: &str) -> String {
+    // Serializing to a String only fails if the Serialize impl decides to fail, which the Serialize impl of `str` never does.
+    serde_json::to_string(s).unwrap()
 }
 
 fn run() -> Result<(), CliError> {
@@ -721,16 +76,14 @@ fn run() -> Result<(), CliError> {
     let mut input = None;
     let mut output = None;
     let mut map = None;
-    let for_browser = true;
-    let mut for_bower = false;
-    let mut es6_syntax = true;
-    let mut es6_syntax_everywhere = true;
+    let mut package_manager = PackageManager::Npm;
     let mut map_inline = false;
     let mut no_map = false;
     let mut watch = false;
     let mut quiet_watch = false;
     let mut external = FnvHashSet::default();
 
+    // TODO replace this arg parser
     let mut iter = opts::args();
     while let Some(arg) = iter.next() {
         let opt = match arg {
@@ -756,12 +109,7 @@ fn run() -> Result<(), CliError> {
             }
             "-I" | "--map-inline" => map_inline = true,
             "-M" | "--no-map" => no_map = true,
-            "-b" | "--for-bower" => for_bower = true,
-            "-e" | "--es-syntax" => es6_syntax = true,
-            "-E" | "--es-syntax-everywhere" => {
-                es6_syntax = true;
-                es6_syntax_everywhere = true;
-            }
+            "-b" | "--for-bower" => package_manager = PackageManager::Bower,
             "-x" | "--external" => {
                 lazy_static! {
                     static ref COMMA: Regex = Regex::new(r#"\s*,\s*"#).unwrap();
@@ -839,10 +187,7 @@ fn run() -> Result<(), CliError> {
     };
 
     let input_options = InputOptions {
-        for_browser,
-        for_bower,
-        es6_syntax,
-        es6_syntax_everywhere,
+        package_manager,
         external,
     };
 
@@ -868,7 +213,7 @@ fn run() -> Result<(), CliError> {
         let ms = elapsed.as_secs() * 1_000 + u64::from(elapsed.subsec_millis());
 
         let (tx, rx) = mpsc::channel();
-        let debounce_dur = time::Duration::from_millis(5);
+        let debounce_duration = time::Duration::from_millis(5);
         let mut watcher = notify::raw_watcher(tx.clone())?;
 
         for path in modules.keys() {
@@ -884,7 +229,7 @@ fn run() -> Result<(), CliError> {
 
         loop {
             let first_event = rx.recv().expect("notify::watcher disconnected");
-            thread::sleep(debounce_dur);
+            thread::sleep(debounce_duration);
             for event in iter::once(first_event).chain(rx.try_iter()) {
                 let _op = event.op?;
             }
@@ -1052,7 +397,7 @@ pub enum CliError {
     Io(io::Error),
     Json(serde_json::Error),
     Notify(notify::Error),
-    Es6(es6::Error),
+    Es6(es6::error::Error),
     Lex(lex::Error),
     ParseStrLit(lex::ParseStrLitError),
     Box(Box<dyn Any + Send + 'static>),
@@ -1072,8 +417,8 @@ impl From<notify::Error> for CliError {
         CliError::Notify(inner)
     }
 }
-impl From<es6::Error> for CliError {
-    fn from(inner: es6::Error) -> CliError {
+impl From<es6::error::Error> for CliError {
+    fn from(inner: es6::error::Error) -> CliError {
         CliError::Es6(inner)
     }
 }
@@ -1163,802 +508,6 @@ fn main() {
             1
         }
     })
-}
-
-trait PathBufExt {
-    fn append_resolving<P: AsRef<Path> + ?Sized>(&mut self, more: &P);
-    fn prepend_resolving<P: AsRef<Path> + ?Sized>(&mut self, base: &P);
-    fn empty(&mut self);
-    fn replace_with<P: AsRef<Path> + ?Sized>(&mut self, that: &P);
-    fn as_mut_vec(&mut self) -> &mut Vec<u8>;
-}
-impl PathBufExt for PathBuf {
-    fn append_resolving<P: AsRef<Path> + ?Sized>(&mut self, more: &P) {
-        for c in more.as_ref().components() {
-            match c {
-                Component::Prefix(prefix) => {
-                    *self = PathBuf::from(prefix.as_os_str().to_owned());
-                }
-                Component::RootDir => {
-                    self.push(path::MAIN_SEPARATOR.to_string());
-                }
-                Component::CurDir => {}
-                Component::ParentDir => {
-                    self.pop();
-                }
-                Component::Normal(part) => {
-                    self.push(part);
-                }
-            }
-        }
-    }
-    fn prepend_resolving<P: AsRef<Path> + ?Sized>(&mut self, base: &P) {
-        let mut tmp = base.as_ref().to_owned();
-        mem::swap(self, &mut tmp);
-        self.append_resolving(tmp.as_path());
-    }
-    fn empty(&mut self) {
-        self.as_mut_vec().clear();
-    }
-    fn replace_with<P: AsRef<Path> + ?Sized>(&mut self, that: &P) {
-        self.empty();
-        self.push(that);
-    }
-
-    fn as_mut_vec(&mut self) -> &mut Vec<u8> {
-        unsafe { &mut *(self as *mut PathBuf as *mut Vec<u8>) }
-    }
-}
-
-trait PathExt {
-    fn is_explicitly_relative(&self) -> bool;
-    fn relative_from<P: AsRef<Path> + ?Sized>(&self, base: &P) -> Option<PathBuf>;
-}
-impl PathExt for Path {
-    #[inline]
-    fn is_explicitly_relative(&self) -> bool {
-        matches!(
-            self.components().next(),
-            Some(Component::CurDir) | Some(Component::ParentDir)
-        )
-    }
-    fn relative_from<P: AsRef<Path> + ?Sized>(&self, base: &P) -> Option<PathBuf> {
-        let base = base.as_ref();
-        if self.is_absolute() != base.is_absolute() {
-            if self.is_absolute() {
-                Some(PathBuf::from(self))
-            } else {
-                None
-            }
-        } else {
-            let mut ita = self.components();
-            let mut itb = base.components();
-            let mut comps: Vec<Component> = vec![];
-            loop {
-                match (ita.next(), itb.next()) {
-                    (None, None) => break,
-                    (Some(a), None) => {
-                        comps.push(a);
-                        comps.extend(ita.by_ref());
-                        break;
-                    }
-                    (None, _) => comps.push(Component::ParentDir),
-                    (Some(a), Some(b)) if comps.is_empty() && a == b => (),
-                    (Some(a), Some(b)) if b == Component::CurDir => comps.push(a),
-                    (Some(_), Some(b)) if b == Component::ParentDir => return None,
-                    (Some(a), Some(_)) => {
-                        comps.push(Component::ParentDir);
-                        for _ in itb {
-                            comps.push(Component::ParentDir);
-                        }
-                        comps.push(a);
-                        comps.extend(ita.by_ref());
-                        break;
-                    }
-                }
-            }
-            Some(comps.iter().map(|c| c.as_os_str()).collect())
-        }
-    }
-}
-
-impl WorkerInit {
-    fn add_work(&self, work: Work) {
-        self.queue.push(work);
-    }
-}
-
-impl Worker {
-    fn new(init: WorkerInit) -> Self {
-        Worker {
-            tx: init.tx,
-            resolver: Resolver::new(init.input_options),
-            queue: init.queue,
-            quit: init.quit,
-        }
-    }
-
-    fn run(mut self) {
-        while let Some(work) = self.get_work() {
-            let work_done = match work {
-                Work::Resolve { context, name } => {
-                    self.resolver
-                        .resolve(&context, &name)
-                        .map(|resolved| WorkDone::Resolve {
-                            context,
-                            name,
-                            resolved,
-                        })
-                }
-                Work::Include { module } => self
-                    .include(&module)
-                    .map(|info| WorkDone::Include { module, info }),
-            };
-            if self.tx.send(work_done).is_err() {
-                return;
-            }
-        }
-    }
-
-    fn include(&self, module: &Path) -> Result<ModuleInfo, CliError> {
-        let source = {
-            let file = fs::File::open(module)?;
-            let mut buf_reader = io::BufReader::new(file);
-            let mut bytes = Vec::new();
-            buf_reader.read_to_end(&mut bytes)?;
-            match String::from_utf8(bytes) {
-                Ok(s) => s,
-                Err(err) => {
-                    return Err(CliError::InvalidUtf8 {
-                        context: module.to_owned(),
-                        err,
-                    });
-                }
-            }
-        };
-        let mut new_source = None;
-        let prefix;
-        let suffix;
-
-        let deps = {
-            let path_string = module.to_string_lossy();
-            // module.to_str().ok_or("<path with invalid utf-8>")
-            let mut lexer = lex::Lexer::new(path_string.as_ref(), &source);
-
-            let deps;
-            let ext = module.extension();
-            if matches!(ext, Some(s) if s == "mjs") {
-                let module = es6::module_to_cjs(&mut lexer, false)?;
-                // println!("{:#?}", module);
-                deps = module.deps;
-                prefix = module.source_prefix;
-                suffix = module.source_suffix;
-                new_source = Some(module.source);
-            } else if matches!(ext, Some(s) if s == "json") {
-                deps = FnvHashSet::default();
-                prefix = "module.exports =".to_owned();
-                suffix = String::new();
-            } else if self.resolver.input_options.es6_syntax_everywhere {
-                let module = es6::module_to_cjs(&mut lexer, true)?;
-                // println!("{:#?}", module);
-                deps = module.deps;
-                prefix = module.source_prefix;
-                suffix = module.source_suffix;
-                new_source = Some(module.source);
-            } else {
-                deps = cjs_parse_deps(&mut lexer)?;
-                prefix = String::new();
-                suffix = String::new();
-            }
-
-            if let Some(error) = lexer.take_error() {
-                return Err(From::from(error));
-            }
-
-            deps.into_iter().map(|s| s.into_owned()).collect()
-        };
-
-        // Convert hashbang #! to //
-        if new_source.as_ref().unwrap_or(&source).starts_with("#!") {
-            if new_source.is_none() {
-                new_source = Some(source.clone())
-            }
-            new_source.as_mut().unwrap().replace_range(0..2, "//")
-        }
-
-        Ok(ModuleInfo {
-            source: match new_source {
-                None => Source {
-                    prefix,
-                    body: source,
-                    suffix,
-                    original: None,
-                },
-                Some(new_source) => Source {
-                    prefix,
-                    body: new_source,
-                    suffix,
-                    original: Some(source),
-                },
-            },
-            deps,
-        })
-    }
-
-    fn get_work(&mut self) -> Option<Work> {
-        loop {
-            match self.queue.try_pop() {
-                Some(work) => return Some(work),
-                None => {
-                    if self.quit.load(Ordering::Relaxed) {
-                        return None;
-                    } else {
-                        thread::yield_now();
-                        // thread::sleep(time::Duration::from_millis(1));
-                    }
-                }
-            }
-        }
-    }
-}
-
-fn get_component_dir(for_bower: bool) -> &'static str {
-    if for_bower {
-        "bower_components"
-    } else {
-        "node_modules"
-    }
-}
-
-impl Resolver {
-    fn new(input_options: InputOptions) -> Self {
-        Resolver {
-            input_options,
-            ..Default::default()
-        }
-    }
-
-    #[inline]
-    fn needs_dir(name: &str, path: &Path) -> bool {
-        name.ends_with('/')
-            || matches!(
-                path.components().last(),
-                Some(Component::CurDir) | Some(Component::ParentDir)
-            )
-    }
-
-    fn resolve_main(&self, mut dir: PathBuf, name: &str) -> Result<Resolved, CliError> {
-        let path = Path::new(name);
-        dir.append_resolving(path);
-        let needs_dir = Self::needs_dir(name, path);
-        self.resolve_path_or_module(None, dir, needs_dir, false)?
-            .ok_or_else(|| CliError::MainNotFound {
-                name: name.to_owned(),
-            })
-    }
-
-    fn resolve(&self, context: &Path, name: &str) -> Result<Resolved, CliError> {
-        if name.is_empty() {
-            return Err(CliError::EmptyModuleName {
-                context: context.to_owned(),
-            });
-        }
-
-        let path = Path::new(name);
-        let needs_dir = Self::needs_dir(name, path);
-        if path.is_absolute() {
-            Ok(self
-                .resolve_path_or_module(Some(context), path.to_owned(), needs_dir, false)?
-                .ok_or_else(|| CliError::ModuleNotFound {
-                    context: context.to_owned(),
-                    name: name.to_owned(),
-                })?)
-        } else if path.is_explicitly_relative() {
-            let mut dir = context.to_owned();
-            let did_pop = dir.pop(); // to directory
-            debug_assert!(did_pop);
-            dir.append_resolving(path);
-            Ok(self
-                .resolve_path_or_module(Some(context), dir, needs_dir, false)?
-                .ok_or_else(|| CliError::ModuleNotFound {
-                    context: context.to_owned(),
-                    name: name.to_owned(),
-                })?)
-        } else {
-            match self.module_substitution(context, name)? {
-                ModuleSubstitution::Ignore => return Ok(Resolved::Ignore),
-                ModuleSubstitution::External => return Ok(Resolved::External),
-                ModuleSubstitution::Replace(new_name) => {
-                    // TODO: detect cycles
-                    // eprintln!("module replace {} => {}", name, &new_name);
-                    return self.resolve(context, &new_name);
-                }
-                ModuleSubstitution::Normal => {}
-            }
-            let component_dir = get_component_dir(self.input_options.for_bower);
-
-            let mut suffix = PathBuf::from(component_dir);
-            suffix.push(name);
-
-            let mut dir = context.to_owned();
-            while dir.pop() {
-                match dir.file_name() {
-                    Some(s) if s == component_dir => continue,
-                    _ => {}
-                }
-                let new_path = dir.join(&suffix);
-                if let Some(result) =
-                    self.resolve_path_or_module(Some(context), new_path, needs_dir, false)?
-                {
-                    return Ok(result);
-                }
-            }
-
-            Err(CliError::ModuleNotFound {
-                context: context.to_owned(),
-                name: name.to_owned(),
-            })
-        }
-    }
-
-    fn module_substitution(
-        &self,
-        context: &Path,
-        name: &str,
-    ) -> Result<ModuleSubstitution, CliError> {
-        let module_name = name.split('/').next().unwrap();
-        if self.input_options.external.contains(module_name) {
-            return Ok(ModuleSubstitution::External);
-        }
-        if self.input_options.for_browser {
-            if let Some(p) = context.parent() {
-                if let Some(info) = self
-                    .cache
-                    .nearest_package_info(p.to_owned(), self.input_options.for_bower)?
-                {
-                    match info.browser_substitutions.0.get(Path::new(module_name)) {
-                        Some(&BrowserSubstitution::Ignore) => {
-                            return Ok(ModuleSubstitution::Ignore)
-                        }
-                        Some(&BrowserSubstitution::Replace(ref to)) => {
-                            let mut new_name = to.to_string_lossy().into_owned();
-                            new_name.push_str(&name[module_name.len()..]);
-                            return Ok(ModuleSubstitution::Replace(new_name));
-                        }
-                        None => {}
-                    }
-                }
-            }
-        }
-        Ok(ModuleSubstitution::Normal)
-    }
-
-    fn resolve_path_or_module(
-        &self,
-        context: Option<&Path>,
-        mut path: PathBuf,
-        needs_dir: bool,
-        package: bool,
-    ) -> Result<Option<Resolved>, CliError> {
-        let package_info = if self.input_options.for_browser {
-            self.cache
-                .nearest_package_info(path.clone(), self.input_options.for_bower)?
-        } else {
-            None
-        };
-
-        macro_rules! check_path {
-            ( $package_info:ident, $path:ident ) => {
-                // eprintln!("check {}", $path.display());
-                if self.input_options.for_browser {
-                    match Self::check_path($package_info.as_ref().map(|x| x.as_ref()), &$path) {
-                        PathSubstitution::Normal => {
-                            // eprintln!("resolve {}", $path.display());
-                            return Ok(Some(Resolved::Normal($path)));
-                        }
-                        PathSubstitution::Ignore => return Ok(Some(Resolved::Ignore)),
-                        PathSubstitution::Replace(p) => {
-                            // eprintln!("path replace {} => {}", $path.display(), p.display());
-                            return Ok(Some(Resolved::Normal(p)));
-                        }
-                        PathSubstitution::Missing => {}
-                    }
-                } else if path.is_file() {
-                    return Ok(Some(Resolved::Normal(path)));
-                }
-            };
-        }
-
-        if !needs_dir {
-            // <path>
-            check_path!(package_info, path);
-
-            let file_name = path
-                .file_name()
-                .ok_or_else(|| CliError::RequireRoot {
-                    context: context.map(|p| p.to_owned()),
-                    path: path.clone(),
-                })?
-                .to_owned();
-
-            let mut new_file_name = file_name.to_owned();
-
-            if self.input_options.es6_syntax {
-                // <path>.mjs
-                new_file_name.push(".mjs");
-                path.set_file_name(&new_file_name);
-                check_path!(package_info, path);
-                new_file_name.clear();
-                new_file_name.push(&file_name);
-            }
-
-            // <path>.js
-            new_file_name.push(".js");
-            path.set_file_name(&new_file_name);
-            check_path!(package_info, path);
-
-            // <path>.json
-            new_file_name.push("on"); // .js|on
-            path.set_file_name(&new_file_name);
-            check_path!(package_info, path);
-
-            path.set_file_name(&file_name);
-        }
-
-        if !package {
-            if let Some(info) = self
-                .cache
-                .package_info(&mut path, self.input_options.for_bower)?
-            {
-                path.replace_with(&info.main);
-                return self.resolve_path_or_module(context, path, false, true);
-            }
-        }
-
-        if self.input_options.es6_syntax {
-            // <path>/index.mjs
-            path.push("index.mjs");
-            check_path!(package_info, path);
-            path.pop();
-        }
-
-        // <path>/index.js
-        path.push("index.js");
-        check_path!(package_info, path);
-        path.pop();
-
-        // <path>/index.json
-        path.push("index.json");
-        check_path!(package_info, path);
-        // path.pop();
-
-        Ok(None)
-    }
-
-    fn check_path(package_info: Option<&PackageInfo>, path: &Path) -> PathSubstitution {
-        if let Some(package_info) = package_info {
-            match package_info.browser_substitutions.0.get(path) {
-                Some(BrowserSubstitution::Ignore) => return PathSubstitution::Ignore,
-                Some(BrowserSubstitution::Replace(ref path)) => {
-                    return PathSubstitution::Replace(path.clone());
-                }
-                None => {}
-            }
-        }
-        if path.is_file() {
-            PathSubstitution::Normal
-        } else {
-            PathSubstitution::Missing
-        }
-    }
-}
-
-impl PackageCache {
-    fn nearest_package_info(
-        &self,
-        mut dir: PathBuf,
-        for_bower: bool,
-    ) -> Result<Option<Rc<PackageInfo>>, CliError> {
-        loop {
-            if !matches!(dir.file_name(), Some(s) if s == get_component_dir(for_bower)) {
-                if let Some(info) = self.package_info(&mut dir, for_bower)? {
-                    return Ok(Some(info));
-                }
-            }
-            if !dir.pop() {
-                return Ok(None);
-            }
-        }
-    }
-    fn package_info(
-        &self,
-        dir: &mut PathBuf,
-        for_bower: bool,
-    ) -> Result<Option<Rc<PackageInfo>>, CliError> {
-        let mut pkgs = self.pkgs.borrow_mut();
-        let manifest_file_name = if for_bower {
-            "bower.json"
-        } else {
-            "package.json"
-        };
-        Ok(pkgs
-            .entry(dir.clone())
-            .or_insert_with(|| {
-                dir.push(manifest_file_name);
-                if let Ok(file) = fs::File::open(&dir) {
-                    let buf_reader = io::BufReader::new(file);
-                    if let Ok(mut info) = serde_json::from_reader::<_, PackageInfo>(buf_reader) {
-                        dir.pop();
-                        info.set_base(&dir);
-                        // eprintln!("info {} {:?}", dir.display(), info);
-                        Some(Rc::new(info))
-                    } else {
-                        dir.pop();
-                        None
-                    }
-                } else {
-                    dir.pop();
-                    None
-                }
-            })
-            .as_ref()
-            .cloned())
-    }
-}
-
-#[derive(Debug, Default, PartialEq, Eq, Clone)]
-struct PackageInfo {
-    main: PathBuf,
-    browser_substitutions: BrowserSubstitutionMap,
-}
-impl PackageInfo {
-    fn set_base(&mut self, base: &Path) {
-        self.main.prepend_resolving(base);
-        let substs = mem::replace(&mut self.browser_substitutions, Default::default());
-        self.browser_substitutions
-            .0
-            .extend(substs.0.into_iter().map(|(mut from, mut to)| {
-                if from.is_explicitly_relative() {
-                    from.prepend_resolving(base);
-                }
-                match to {
-                    BrowserSubstitution::Ignore => {}
-                    BrowserSubstitution::Replace(ref mut path) => {
-                        if path.is_explicitly_relative() {
-                            path.prepend_resolving(base);
-                        }
-                    }
-                }
-                (from, to)
-            }));
-    }
-}
-impl<'de> Deserialize<'de> for PackageInfo {
-    fn deserialize<D: Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
-        #[derive(Debug, Default, PartialEq, Eq, Deserialize)]
-        #[serde(default)]
-        struct RawPackageInfo {
-            #[serde(deserialize_with = "from_main")]
-            main: Option<PathBuf>,
-            browser: BrowserField,
-        }
-
-        let info = RawPackageInfo::deserialize(deserializer)?;
-        let main = info.main.unwrap_or(PathBuf::from("./index"));
-        let browser_substitutions = info.browser.to_map(&main);
-        Ok(PackageInfo {
-            main,
-            browser_substitutions,
-        })
-    }
-}
-
-macro_rules! visit_unconditionally {
-    // bool i64 i128 u64 u128 f64 str bytes none some unit newtype_struct seq map enum
-    ($l:lifetime $as:expr) => {};
-    ($l:lifetime $as:expr, ) => {};
-    ($l:lifetime $as:expr, bool $($x:tt)*) => {
-        fn visit_bool<E: de::Error>(self, _: bool) -> Result<Self::Value, E> { Ok($as) }
-        visit_unconditionally!($l $as, $($x)*);
-    };
-    ($l:lifetime $as:expr, i64 $($x:tt)*) => {
-        fn visit_i64<E: de::Error>(self, _: i64) -> Result<Self::Value, E> { Ok($as) }
-        visit_unconditionally!($l $as, $($x)*);
-    };
-    ($l:lifetime $as:expr, i128 $($x:tt)*) => {
-        fn visit_i128<E: de::Error>(self, _: i128) -> Result<Self::Value, E> { Ok($as) }
-        visit_unconditionally!($l $as, $($x)*);
-    };
-    ($l:lifetime $as:expr, u64 $($x:tt)*) => {
-        fn visit_u64<E: de::Error>(self, _: u64) -> Result<Self::Value, E> { Ok($as) }
-        visit_unconditionally!($l $as, $($x)*);
-    };
-    ($l:lifetime $as:expr, u128 $($x:tt)*) => {
-        fn visit_u128<E: de::Error>(self, _: u128) -> Result<Self::Value, E> { Ok($as) }
-        visit_unconditionally!($l $as, $($x)*);
-    };
-    ($l:lifetime $as:expr, f64 $($x:tt)*) => {
-        fn visit_f64<E: de::Error>(self, _: f64) -> Result<Self::Value, E> { Ok($as) }
-        visit_unconditionally!($l $as, $($x)*);
-    };
-    ($l:lifetime $as:expr, str $($x:tt)*) => {
-        fn visit_str<E: de::Error>(self, v: &str) -> Result<Self::Value, E> { Ok($as) }
-        visit_unconditionally!($l $as, $($x)*);
-    };
-    ($l:lifetime $as:expr, bytes $($x:tt)*) => {
-        fn visit_bytes<E: de::Error>(self, _: &[u8]) -> Result<Self::Value, E> { Ok($as) }
-        visit_unconditionally!($l $as, $($x)*);
-    };
-    ($l:lifetime $as:expr, none $($x:tt)*) => {
-        fn visit_none<E: de::Error>(self) -> Result<Self::Value, E> { Ok($as) }
-        visit_unconditionally!($l $as, $($x)*);
-    };
-    ($l:lifetime $as:expr, some $($x:tt)*) => {
-        fn visit_some<D: Deserializer<$l>>(self, _: D) -> Result<Self::Value, D::Error> { Ok($as) }
-        visit_unconditionally!($l $as, $($x)*);
-    };
-    ($l:lifetime $as:expr, unit $($x:tt)*) => {
-        fn visit_unit<E: de::Error>(self) -> Result<Self::Value, E> { Ok($as) }
-        visit_unconditionally!($l $as, $($x)*);
-    };
-    ($l:lifetime $as:expr, newtype_struct $($x:tt)*) => {
-        fn visit_newtype_struct<D: Deserializer<$l>>(self, _: D) -> Result<Self::Value, D::Error> { Ok($as) }
-        visit_unconditionally!($l $as, $($x)*);
-    };
-    ($l:lifetime $as:expr, seq $($x:tt)*) => {
-        fn visit_seq<A: de::SeqAccess<$l>>(self, _: A) -> Result<Self::Value, A::Error> { Ok($as) }
-        visit_unconditionally!($l $as, $($x)*);
-    };
-    ($l:lifetime $as:expr, map $($x:tt)*) => {
-        fn visit_map<A: de::MapAccess<$l>>(self, _: A) -> Result<Self::Value, A::Error> { Ok($as) }
-        visit_unconditionally!($l $as, $($x)*);
-    };
-    ($l:lifetime $as:expr, enum $($x:tt)*) => {
-        fn visit_enum<A: de::EnumAccess<$l>>(self, _: A) -> Result<Self::Value, A::Error> { Ok($as) }
-        visit_unconditionally!($l $as, $($x)*);
-    };
-}
-
-#[derive(Debug, PartialEq, Eq, Clone)]
-enum BrowserField {
-    Empty,
-    Main(PathBuf),
-    Complex(BrowserSubstitutionMap),
-}
-impl BrowserField {
-    fn to_map(self, main: &Path) -> BrowserSubstitutionMap {
-        match self {
-            BrowserField::Empty => Default::default(),
-            BrowserField::Main(mut to) => {
-                if !to.is_explicitly_relative() {
-                    to.prepend_resolving(Path::new("."));
-                }
-                BrowserSubstitutionMap(map! {
-                    main.to_owned() => BrowserSubstitution::Replace(to),
-                })
-            }
-            BrowserField::Complex(map) => map,
-        }
-    }
-}
-impl Default for BrowserField {
-    fn default() -> Self {
-        BrowserField::Empty
-    }
-}
-impl<'de> Deserialize<'de> for BrowserField {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        struct BrowserFieldVisitor;
-
-        impl<'de> Visitor<'de> for BrowserFieldVisitor {
-            type Value = BrowserField;
-
-            fn expecting(&self, f: &mut fmt::Formatter) -> fmt::Result {
-                write!(f, "anything at all")
-            }
-            fn visit_map<A: de::MapAccess<'de>>(self, access: A) -> Result<Self::Value, A::Error> {
-                Ok(BrowserField::Complex(Deserialize::deserialize(
-                    de::value::MapAccessDeserializer::new(access),
-                )?))
-            }
-            fn visit_str<E: de::Error>(self, v: &str) -> Result<Self::Value, E> {
-                Ok(BrowserField::Main(PathBuf::from(v)))
-            }
-
-            visit_unconditionally!('de BrowserField::Empty, bool i64 i128 u64 u128 f64 bytes none some unit newtype_struct seq enum);
-        }
-
-        deserializer.deserialize_any(BrowserFieldVisitor)
-    }
-}
-
-// The main might be an array or a string of arrays
-fn from_main<'de, T, D>(deserializer: D) -> Result<Option<T>, D::Error>
-where
-    for<'a> T: From<&'a str> + Deserialize<'de>,
-    D: Deserializer<'de>,
-{
-    struct FromMain<T>(PhantomData<T>);
-
-    impl<'de, T> Visitor<'de> for FromMain<T>
-    where
-        for<'a> T: From<&'a str> + Deserialize<'de>,
-    {
-        type Value = Option<T>;
-
-        fn expecting(&self, f: &mut fmt::Formatter) -> fmt::Result {
-            write!(f, "string or array or nothing")
-        }
-
-        fn visit_str<E: de::Error>(self, v: &str) -> Result<Self::Value, E> {
-            Ok(Some(T::from(v)))
-        }
-
-        fn visit_seq<S>(self, mut seq: S) -> Result<Self::Value, S::Error>
-        where
-            S: SeqAccess<'de>,
-        {
-            while let Ok(item) = seq.next_element() {
-                if let Some(item) = item {
-                    let item: String = item;
-                    if item.ends_with(".js") {
-                        return Ok(Some(T::from(&item)));
-                    }
-                }
-            }
-            Ok(None)
-        }
-
-        visit_unconditionally!('de None, bool i64 i128 u64 u128 f64 bytes none some unit newtype_struct map enum);
-    }
-
-    deserializer.deserialize_any(FromMain(PhantomData))
-}
-
-#[derive(Debug, PartialEq, Eq, Clone)]
-enum BrowserSubstitution<T> {
-    Ignore,
-    Replace(T),
-}
-#[derive(Debug, Default, Deserialize, PartialEq, Eq, Clone)]
-#[serde(transparent)]
-struct BrowserSubstitutionMap(FnvHashMap<PathBuf, BrowserSubstitution<PathBuf>>);
-
-impl<'de, T> Deserialize<'de> for BrowserSubstitution<T>
-where
-    for<'a> T: From<&'a str>,
-{
-    fn deserialize<D: Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
-        struct SubstitutionVisitor<T>(PhantomData<T>);
-
-        impl<'de, T> Visitor<'de> for SubstitutionVisitor<T>
-        where
-            for<'a> T: From<&'a str>,
-        {
-            type Value = BrowserSubstitution<T>;
-
-            fn expecting(&self, f: &mut fmt::Formatter) -> fmt::Result {
-                write!(f, "substitution path or false")
-            }
-
-            fn visit_bool<E: de::Error>(self, v: bool) -> Result<Self::Value, E> {
-                if v {
-                    Err(de::Error::invalid_value(de::Unexpected::Bool(v), &self))
-                } else {
-                    Ok(BrowserSubstitution::Ignore)
-                }
-            }
-
-            fn visit_str<E: de::Error>(self, v: &str) -> Result<Self::Value, E> {
-                Ok(BrowserSubstitution::Replace(From::from(v)))
-            }
-        }
-
-        deserializer.deserialize_any(SubstitutionVisitor(PhantomData))
-    }
 }
 
 #[cfg(test)]
