+++ src/source_maps.rs
@@ -1,0 +1,8 @@
+use std::path::{Path, PathBuf};
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum SourceMapOutput<'a> {
+    Suppressed,
+    Inline,
+    File(PathBuf, &'a Path),
+}
